added some AI move evaluation

+++ Colonecon/GameLogic/Factions/NPCAI/INPCAction.cs
@@ -1,6 +1,6 @@
 public interface INPCAction
 {
-    int Value { get; set; }
+    double Value { get; set; }
     Faction Faction{ get; }
     void ExecuteAction();
 }
